Keep rig run tests under audit limits
