+++ src/commands/runs.rs
@@ -1,3 +1,5 @@
+#[cfg(test)]
+mod artifact_index_tests;
 mod bench;
 mod bundle;
 #[cfg(test)]
@@ -669,86 +671,6 @@ mod tests {
             };
             assert_eq!(output.runs.len(), 1);
             assert_eq!(output.runs[0].id, bench.id);
-        });
-    }
-
-    #[test]
-    fn rig_runs_list_surfaces_compact_artifact_index() {
-        with_isolated_home(|home| {
-            let _xdg = XdgGuard::unset();
-            let store = ObservationStore::open_initialized().expect("store");
-            let run = store
-                .start_run(sample_run(
-                    "rig",
-                    "homeboy",
-                    "proof-rig",
-                    serde_json::json!({
-                        "pipeline": {
-                            "name": "check",
-                            "steps": [
-                                {
-                                    "kind": "command",
-                                    "label": "produce proof report",
-                                    "status": "fail",
-                                    "error": "report failed"
-                                }
-                            ],
-                            "passed": 0,
-                            "failed": 1
-                        }
-                    }),
-                ))
-                .expect("rig run");
-            store
-                .finish_run(&run.id, RunStatus::Fail, None)
-                .expect("finish rig run");
-            let report = home.path().join("proof-report.json");
-            std::fs::write(&report, "{}").expect("report artifact");
-            store
-                .record_artifact(&run.id, "proof_report", &report)
-                .expect("record report");
-
-            let (output, _) = list_runs(
-                RunsListArgs {
-                    runner: None,
-                    kind: None,
-                    component_id: None,
-                    rig: Some("proof-rig".to_string()),
-                    status: None,
-                    limit: 20,
-                },
-                "rig.runs",
-            )
-            .expect("rig runs");
-
-            let RunsOutput::List(output) = output else {
-                panic!("expected list output");
-            };
-            assert_eq!(output.command, "rig.runs");
-            assert_eq!(output.runs.len(), 1);
-            let artifact_index = output.runs[0]
-                .artifact_index
-                .as_ref()
-                .expect("artifact index");
-            assert_eq!(artifact_index.run_id, run.id);
-            assert_eq!(artifact_index.rig_id, "proof-rig");
-            assert_eq!(artifact_index.status, "fail");
-            assert!(artifact_index
-                .artifact_index_path
-                .ends_with("rig-artifact-index.json"));
-            assert_eq!(
-                artifact_index.artifacts_command,
-                format!("homeboy runs artifacts {}", run.id)
-            );
-            assert!(artifact_index
-                .key_report_refs
-                .iter()
-                .any(|artifact| artifact.kind == "proof_report"));
-            assert_eq!(artifact_index.failed_step_refs.len(), 1);
-            assert_eq!(
-                artifact_index.failed_step_refs[0].label,
-                "produce proof report"
-            );
         });
     }
